+++ modules/ZWave/index.js
@@ -1,8 +1,8 @@
 /*** Z-Wave Binding module ********************************************************
 
-Version: 2.1.3
+Version: 2.2.0
 -------------------------------------------------------------------------------
-Author: Serguei Poltorak <ps@z-wave.me>
+Author: Serguei Poltorak <ps@z-wave.me>, Niels Roche <nir@zwave.eu>
 Copyright: (c) Z-Wave.Me, 2014
 
 ******************************************************************************/
@@ -315,9 +315,10 @@ ZWave.prototype.CommunicationLogger = function() {
 	uniq = function(arr) { return arr.filter(function(value, index, self) { return self.indexOf(value) === index; }); };
 	group = function(arr) { var ret = {}; arr.map(function(x) { if (ret[x]) ret[x]++; else ret[x] = 1; }); return ret; };
 
+	/*
 	inH = function () {
 
-		console.log(JSON.stringify(this));
+		//console.log(JSON.stringify(this));
 
 		//if(ipacket.length > 4000) { ipacket = [];}
 
@@ -334,22 +335,22 @@ ZWave.prototype.CommunicationLogger = function() {
 				return x.RSSI.value;
 			});
 			console.log("homeid: "+id, "avg rssi: "+avg(rssis), stddev(rssis), "rssi length: "+rssis.length);
-		};*/
+		};
 
-		saveObject("incomingPacket.json", ipacket);
+		//saveObject("incomingPacket.json", ipacket);
 	};
 
 	//zway.controller.data.incomingPacket.bind(inH);
 
 	outH = function () {
 
-		console.log(JSON.stringify(this));
+		//console.log(JSON.stringify(this));
 
 		//if(opacket.length > 4000) { opacket = [];}
 
 		opacket.push(this);
 
-		/*
+
 		outNodes = uniq(opacket.filter(function(x) { return x.nodeId }).map(function(x) { return x.nodeId.value }));
 
 		console.log("id PER delivery time, RSSI, hops");
@@ -382,11 +383,12 @@ ZWave.prototype.CommunicationLogger = function() {
 
 			console.log("homeid: "+id, "delivered: "+avg(delivered).toFixed(2), "deliveryTime: "+avg(deliveryTime).toFixed(2), "delivered: "+stddev(deliveryTime).toFixed(2), "avg rssi: "+avg(rssis).toFixed(2), stddev(rssis).toFixed(2), "avg hops: "+avg(hops).toFixed(2));
 		};
-		*/
-		saveObject("outgoingPacket.json", opacket);
+
+		//saveObject("outgoingPacket.json", opacket);
 	};
 
 	//zway.controller.data.outgoingPacket.bind(outH);
+	*/
 
 	/*
 	rssiH = function() {
@@ -452,8 +454,8 @@ ZWave.prototype.externalAPIAllow = function (name) {
 	ws.allowExternalAccess(_name + ".PostfixUpdate", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
 	ws.allowExternalAccess(_name + ".Postfix", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
 	ws.allowExternalAccess(_name + ".PostfixAdd", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
-    ws.allowExternalAccess(_name + ".PostfixGet", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
-    ws.allowExternalAccess(_name + ".PostfixRemove", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
+	ws.allowExternalAccess(_name + ".PostfixGet", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
+	ws.allowExternalAccess(_name + ".PostfixRemove", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
 	ws.allowExternalAccess(_name + ".ExpertConfigGet", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
 	ws.allowExternalAccess(_name + ".ExpertConfigUpdate", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
 	// -- see below -- // ws.allowExternalAccess(_name + ".JSONtoXML", this.config.publicAPI ? this.controller.auth.ROLE.ANONYMOUS : this.controller.auth.ROLE.ADMIN);
@@ -478,8 +480,8 @@ ZWave.prototype.externalAPIRevoke = function (name) {
 	ws.revokeExternalAccess(_name + ".PostfixUpdate");
 	ws.revokeExternalAccess(_name + ".Postfix");
 	ws.revokeExternalAccess(_name + ".PostfixAdd");
-    ws.revokeExternalAccess(_name + ".PostfixGet");
-    ws.revokeExternalAccess(_name + ".PostfixRemove");
+	ws.revokeExternalAccess(_name + ".PostfixGet");
+	ws.revokeExternalAccess(_name + ".PostfixRemove");
 	ws.revokeExternalAccess(_name + ".ExpertConfigGet");
 	ws.revokeExternalAccess(_name + ".ExpertConfigUpdate");
 	// -- see below -- // ws.revokeExternalAccess(_name + ".JSONtoXML");
@@ -884,7 +886,7 @@ ZWave.prototype.defineHandlers = function () {
 			var filterObj = null;
 		}
 
-		ipacket.forEach(function (packet) {
+		/*ipacket.forEach(function (packet) {
 			var exist = _.find(packets, function(p){
 				if(p.updateTime === packet.updateTime && p.type === 'incoming') {
 					if(_.isArray(p.value) && _.isArray(packet.value)) {
@@ -910,9 +912,10 @@ ZWave.prototype.defineHandlers = function () {
 				);
 				console.log("incomming:" +JSON.stringify(packets));
 			}
-		});
 
-		opacket.forEach(function (packet) {
+		});*/
+
+		/*opacket.forEach(function (packet) {
 			var exist = _.find(packets, function(p){
 				if(p.updateTime === packet.updateTime && p.type === 'outgoing') {
 					if(_.isArray(p.value) && _.isArray(packet.value)) {
@@ -924,7 +927,7 @@ ZWave.prototype.defineHandlers = function () {
 			});
 
 			if(typeof exist === 'undefined') {
-				console.log("outgoing:" +JSON.stringify(packets));
+				//console.log("outgoing:" +JSON.stringify(packets));
 				packets.push(
 					{
 						type: 'outgoing',
@@ -936,14 +939,14 @@ ZWave.prototype.defineHandlers = function () {
 						application: (_.isArray(packet.value)) ? packetApplication(packet.value) : ""
 					}
 				);
-				console.log("outgoing:" +JSON.stringify(packets));
+				//console.log("outgoing:" +JSON.stringify(packets));
 			}
-		});
+		});*/
 
 		function packetApplication(packet) {
 			var cmdClassKey = decToHex(packet[5], 2, '0x');
 
-			var cmdKey = decToHex(packet[6], 2, '0x')
+			var cmdKey = decToHex(packet[6], 2, '0x');
 
 			var ret = {};
 
@@ -1085,7 +1088,14 @@ ZWave.prototype.defineHandlers = function () {
 
 			if (data.file && data.file.content) {
 				// update firmware from file
-				fwUpdate.Perform(manufacturerId, firmwareId, targetId, data.file.content);
+				var fw;
+				if (data.file.content.substr(0, 1) === ":") {
+					// this is a .hex file
+					fw = IntelHex2bin(data.file.content);
+				} else {
+					fw = data.file.content;
+				}
+				fwUpdate.Perform(manufacturerId, firmwareId, targetId, fw);
 			} else if (data.url) {
 				// update firmware from url
 				http.request({
@@ -1094,7 +1104,14 @@ ZWave.prototype.defineHandlers = function () {
 					contentType: "application/octet-stream", // enforce binary response
 					async: true,
 					success: function (res) {
-						fwUpdate.Perform(manufacturerId, firmwareId, targetId, res.data);
+						var fw;
+						if (res.data.substr(0, 1) === ":") {
+							// this is a .hex file
+							fw = IntelHex2bin(res.data);
+						} else {
+							fw = res.data;
+						}
+						fwUpdate.Perform(manufacturerId, firmwareId, targetId, fw);
 					},
 					error: function (res) {
 						console.error("Failed to download firmware: " + res.statusText);
@@ -2013,25 +2030,54 @@ ZWave.prototype.gateDevicesStart = function () {
 													case 'icon':
 														if (splittedEntry[1] && splittedEntry[1].indexOf(devICC) > -1 && c.data.lastIncludedDevice.value === nodeId) {
 															//add devId
-															if (!changeVDev[splittedEntry[1]]) {
-																changeVDev[splittedEntry[1]] = {};
+															var nId = nodeId + '-' + splittedEntry[1];
+
+															if (!changeVDev[nId]) {
+																changeVDev[nId] = {};
 															}
 
 															// devId (instId-CC-sCC-eT) / postFix type / value - fallback true for hide / deactivate
-															changeVDev[splittedEntry[1]][splittedEntry[0]] = splittedEntry[2] ? splittedEntry[2] : true;
+															changeVDev[nId][splittedEntry[0]] = splittedEntry[2] ? splittedEntry[2] : true;
+														}
+
+														break;
+													case 'discreteState':
+														if (splittedEntry[1] && splittedEntry[1].indexOf(devICC) > -1 && c.data.lastIncludedDevice.value === nodeId) {
+															//add devId
+															var nId = nodeId + '-' + splittedEntry[1];
+
+															if (!changeVDev[nId]) {
+																changeVDev[nId] = {};
+															}
+
+															if (!changeVDev[nId]['discreteState']) {
+																changeVDev[nId]['discreteState'] = {};
+															}
+
+															// devId (instId-CC-sCC-eT) / postFix type / scene + keyAttribute / value - fallback undefined
+															changeVDev[nId]['discreteState'][splittedEntry[2]] = {
+																cnt: splittedEntry[3] ? splittedEntry[3] : undefined,
+																action: splittedEntry[4] ? splittedEntry[4] : undefined,
+																type: splittedEntry[5] ? splittedEntry[5] : undefined
+															};
+															//console.log('discreteState',splittedEntry[3]);
+															//console.log('changeVDev',JSON.stringify(changeVDev, null, 4));
 														}
 
 														break;
 													case 'noVDev':
-
 														if (splittedEntry[1] && splittedEntry[1].indexOf(devICC) > -1) {
+
+															var nId = nodeId + '-' + splittedEntry[1];
+
 															//add devId
-															if (!changeVDev[splittedEntry[1]]) {
-																changeVDev[splittedEntry[1]] = {};
+															//add devId
+															if (!changeVDev[nId]) {
+																changeVDev[nId] = {};
 															}
 
 															// devId (instId-CC-sCC-eT) without creation
-															changeVDev[splittedEntry[1]].noVDev = true;
+															changeVDev[nId].noVDev = true;
 														}
 
 														break;
@@ -2059,7 +2105,7 @@ ZWave.prototype.gateDevicesStart = function () {
 						}
 					}
 
-					var ccId = instanceId + '-' + commandClassId;
+					var ccId = nodeId + '-' + instanceId + '-' + commandClassId;
 
 					if (!changeVDev[ccId] || (changeVDev[ccId] && !changeVDev[ccId].noVDev)) {
 						self.parseAddCommandClass(nodeId, instanceId, commandClassId, false, changeVDev);
@@ -2132,7 +2178,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 		vDevIdNI = nodeId + separ + instanceId,
 		vDevIdC = commandClassId,
 		vDevId = vDevIdPrefix + vDevIdNI + separ + vDevIdC,
-		changeDevId = instanceId + separ + vDevIdC,
+		changeDevId = vDevIdNI + separ + vDevIdC,
 		defaults;
 		// vDev is not in this scope, but in {} scope for each type of device to allow reuse it without closures
 
@@ -2162,7 +2208,9 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 			var args = [],
 				sortArgs = [],
 				last = 0,
-				addVendor = true;
+				addVendor = true,
+				lastId = '',
+				lastIdArr = [];
 
 			for (var i = 0; i < arguments.length; i++) {
 				args.push(arguments[i]);
@@ -2197,7 +2245,22 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 			}
 
 			// add id
-			sortArgs.push('(' + args[last].replace(/-/g, '.') + ')');
+			lastIdArr = args[last].split('-');
+
+			// devices[nodeId].instances[0].commandClasses[96]
+			if (this.zway.devices[lastIdArr[0]].instances[0].commandClasses[96] && Object.keys(this.zway.devices[lastIdArr[0]].instances).length > 1) {
+				lastId = '(' + args[last].replace(/-/g, '.') + ')';
+			} else {
+				lastId = '(#' + lastIdArr[0] + ')';
+			}
+
+			/*if (args[last].indexOf('-0') > -1 ) {
+				lastId = args[last].split('-').shift();
+			} else {
+				lastId = '(' + args[last].replace(/-/g, '.') + ')';
+			}*/
+
+			sortArgs.push(lastId);
 			
 			return sortArgs.join(' ');
 		}
@@ -2229,6 +2292,10 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 			defaultObj.metrics.title = changeObj.rename? compileTitle(changeObj.rename, devIdNI, false) : defaultObj.metrics.title;
 			defaultObj.visibility = changeObj.hide? false : true;
 			defaultObj.permanently_hidden = changeObj.deactivate? true : false;
+
+			if (defaultObj.metrics.discreteStates) {
+				defaultObj.metrics.discreteStates = changeObj.discreteState && !_.isEmpty(changeObj.discreteState)? changeObj.discreteState : defaultObj.metrics.discreteStates;
+			}
 
 			postfixLog(devId, changeObj);
 
@@ -2383,7 +2450,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 					probeType: 'switchColor_rgb',
 					metrics: {
 						icon: 'multilevel',
-						title: compileTitle('Color', vDevIdNI + separ + vDevIdC),
+						title: compileTitle('Color', vDevIdNI),
 						color: {r: cc.data[COLOR_RED].level.value, g: cc.data[COLOR_GREEN].level.value, b: cc.data[COLOR_BLUE].level.value},
 						level: 'off'
 					}
@@ -2391,7 +2458,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 
 				// apply postfix if available
 				if (changeVDev[changeDevId]) {
-					defaults = applyPostfix(defaults, changeVDev[changeDevId], vDevId + separ + "rgb", vDevIdNI + separ + vDevIdC);
+					defaults = applyPostfix(defaults, changeVDev[changeDevId], vDevId + separ + "rgb", vDevIdNI);
 				}
 
 				var vDev_rgb = this.controller.devices.create({
@@ -2444,14 +2511,14 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 							probeType: '',
 							metrics: {
 								icon: 'multilevel',
-								title: compileTitle(cc.data[colorId].capabilityString.value, vDevIdNI + separ + vDevIdC + separ + colorId),
+								title: compileTitle(cc.data[colorId].capabilityString.value, vDevIdNI),
 								level: 'off'
 							}
 						}
 
 						// apply postfix if available
 						if (changeVDev[cVDId]) {
-							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + colorId, vDevIdNI + separ + vDevIdC + separ + colorId);
+							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + colorId, vDevIdNI);
 						}
 
 						switch(colorId) {
@@ -2570,7 +2637,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 					if (!changeVDev[cVDId] || changeVDev[cVDId] && !changeVDev[cVDId].noVDev) {
 
 						defaults.metrics.probeTitle = cc.data[sensorTypeId].sensorTypeString.value;
-						defaults.metrics.title = compileTitle('Sensor', defaults.metrics.probeTitle, vDevIdNI + separ + vDevIdC + separ + sensorTypeId);
+						defaults.metrics.title = compileTitle('Sensor', defaults.metrics.probeTitle, vDevIdNI);
 						// aivs // Motion icon for Sensor Binary by default
 						defaults.metrics.icon = "motion";
 						defaults.probeType = "general_purpose";
@@ -2600,7 +2667,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 
 						// apply postfix if available
 						if (changeVDev[cVDId]) {
-							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + sensorTypeId, vDevIdNI + separ + vDevIdC + separ + sensorTypeId);
+							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + sensorTypeId, vDevIdNI);
 						}
 
 						var vDev = self.controller.devices.create({
@@ -2662,7 +2729,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 
 						defaults.metrics.probeTitle = cc.data[sensorTypeId].sensorTypeString.value;
 						defaults.metrics.scaleTitle = cc.data[sensorTypeId].scaleString.value;
-						defaults.metrics.title = compileTitle('Sensor', defaults.metrics.probeTitle, vDevIdNI + separ + vDevIdC + separ + sensorTypeId);
+						defaults.metrics.title = compileTitle('Sensor', defaults.metrics.probeTitle, vDevIdNI);
 						if (sensorTypeId === 1) {
 							defaults.metrics.icon = "temperature";
 							defaults.probeType = defaults.metrics.icon;
@@ -2697,7 +2764,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 
 						// apply postfix if available
 						if (changeVDev[cVDId]) {
-							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + sensorTypeId, vDevIdNI + separ + vDevIdC + separ + sensorTypeId);
+							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + sensorTypeId, vDevIdNI);
 						}
 
 						var vDev = self.controller.devices.create({
@@ -2756,7 +2823,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 					if (!changeVDev[cVDId] || changeVDev[cVDId] && !changeVDev[cVDId].noVDev) {
 						defaults.metrics.probeTitle = cc.data[scaleId].sensorTypeString.value;
 						defaults.metrics.scaleTitle = cc.data[scaleId].scaleString.value;
-						defaults.metrics.title = compileTitle('Meter', defaults.metrics.probeTitle, vDevIdNI + separ + vDevIdC + separ + scaleId);
+						defaults.metrics.title = compileTitle('Meter', defaults.metrics.probeTitle, vDevIdNI);
 
 						switch (scaleId) {
 							case 0:
@@ -2783,7 +2850,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 
 						// apply postfix if available
 						if (changeVDev[cVDId]) {
-							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + scaleId, vDevIdNI + separ + vDevIdC + separ + scaleId);
+							defaults = applyPostfix(defaults, changeVDev[cVDId], vDevId + separ + scaleId, vDevIdNI);
 						}
 
 						var vDev = self.controller.devices.create({
@@ -3078,7 +3145,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 
 						// check if it should be created
 						if (!changeVDev[cVDId] || changeVDev[cVDId] && !changeVDev[cVDId].noVDev) {
-							a_defaults.metrics.title = compileTitle('Alarm', cc.data[sensorTypeId].typeString.value, vDevIdNI + separ + vDevIdC + separ + sensorTypeId);
+							a_defaults.metrics.title = compileTitle('Alarm', cc.data[sensorTypeId].typeString.value, vDevIdNI);
 
 							switch (sensorTypeId) {
 								case 0:
@@ -3123,7 +3190,7 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 
 							// apply postfix if available
 							if (changeVDev[cVDId]) {
-								a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI + separ + vDevIdC + separ + sensorTypeId);
+								a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI);
 							}
 
 							var a_vDev = self.controller.devices.create({
@@ -3188,12 +3255,12 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 								var a_id = vDevId + separ + notificationTypeId + separ + 'Door' + separ + "A";
 
 								if (!self.controller.devices.get(a_id)) {
-									a_defaults.metrics.title = compileTitle('Alarm', cc.data[notificationTypeId].typeString.value, vDevIdNI + separ + vDevIdC + separ + notificationTypeId + separ + 'Door');
+									a_defaults.metrics.title = compileTitle('Alarm', cc.data[notificationTypeId].typeString.value, vDevIdNI);
 									a_defaults.probeType = 'alarm_door';
 
 									// apply postfix if available
 									if (changeVDev[cVDId]) {
-										a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI + separ + vDevIdC + separ + notificationTypeId + separ + 'Door');
+										a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI);
 									}
 
 									var a_vDev = self.controller.devices.create({
@@ -3304,11 +3371,11 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 											if (!changeVDev[cVDId] || changeVDev[cVDId] && !changeVDev[cVDId].noVDev) {
 
 												if (!self.controller.devices.get(a_id)) {
-													a_defaults.metrics.title = compileTitle('Alarm', cc.data[notificationTypeId].typeString.value, vDevIdNI + separ + vDevIdC + separ + notificationTypeId + separ + eventTypeId);
+													a_defaults.metrics.title = compileTitle('Alarm', cc.data[notificationTypeId].typeString.value, vDevIdNI);
 
 													// apply postfix if available
 													if (changeVDev[cVDId]) {
-														a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI + separ + vDevIdC + separ + notificationTypeId + separ + eventTypeId);
+														a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI);
 													}
 
 													var a_vDev = self.controller.devices.create({
@@ -3354,11 +3421,11 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 									var cVDId = changeDevId + separ + notificationTypeId + separ + eventTypeId;
 									// check if it should be created
 									if (!changeVDev[cVDId] || changeVDev[cVDId] && !changeVDev[cVDId].noVDev) {
-										a_defaults.metrics.title = compileTitle('Alarm', cc.data[notificationTypeId].typeString.value, vDevIdNI + separ + vDevIdC + separ + notificationTypeId + separ + eventTypeId);
+										a_defaults.metrics.title = compileTitle('Alarm', cc.data[notificationTypeId].typeString.value, vDevIdNI);
 
 										// apply postfix if available
 										if (changeVDev[cVDId]) {
-											a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI + separ + vDevIdC + separ + notificationTypeId + separ + eventTypeId);
+											a_defaults = applyPostfix(a_defaults, changeVDev[cVDId], a_id, vDevIdNI);
 										}
 
 										var a_vDev = self.controller.devices.create({
@@ -3399,6 +3466,111 @@ ZWave.prototype.parseAddCommandClass = function (nodeId, instanceId, commandClas
 						self.parseAddCommandClass(nodeId, instanceId, commandClassId, true, changeVDev);
 					}
 				}, "child");
+			}
+		} else if (this.CC["CentralScene"] === commandClassId) {
+
+			var devId = vDevId + separ + 'DS';
+
+			defaults = {
+				deviceType: 'sensorDiscrete',
+				probeType: 'control',
+				metrics: {
+					probeTitle: 'Control',
+					icon: 'gesture',
+					level: '',
+					title: compileTitle('Sensor', 'Control', vDevIdNI),
+                    state: '',
+					/* GESTURES (state):
+					 * hold,
+					 * press / tap (cnt),
+					 * release,
+					 * swipe_up,
+					 * swipe_down,
+					 * swipe_left,
+					 * swipe_right,
+					 * swipe_top_left_to_bottom_right,
+					 * swipe_top_right_to_bottom_left,
+					 * swipe_bottom_left_to_top_right,
+					 * swipe_bottom_right_to_top_left
+					 */
+					currentScene: '',
+					discreteStates: {}
+				}
+			};
+
+			// apply postfix if available
+			if (changeVDev[changeDevId]) {
+				defaults = applyPostfix(defaults, changeVDev[changeDevId], devId, vDevIdNI);
+			}
+
+			var vDev = self.controller.devices.create({
+				deviceId: devId,
+				defaults: defaults,
+				overlay: {},
+				handler: function (command) {
+					if (command === "update") {
+						cc.Get;
+					}
+				},
+				moduleId: self.id
+			});
+
+			if (vDev) {
+				self.dataBind(self.gateDataBinding, self.zway, nodeId, instanceId, commandClassId, "currentScene", function(type) {
+					if (type === self.ZWAY_DATA_CHANGE_TYPE["Deleted"]) {
+						self.controller.devices.remove(devId);
+					} else {
+						try {
+							// output curScene + keyAttr or ''
+							var cS = cc.data['currentScene'].value && !!cc.data['currentScene'].value? cc.data['currentScene'].value : '',
+								mC = cc.data['maxScenes'].value && !!cc.data['maxScenes'].value? cc.data['maxScenes'].value : '',
+								kA = cc.data['keyAttribute'].value && !!cc.data['keyAttribute'].value? cc.data['keyAttribute'].value : '',
+								/*
+								 * CentralScene v3:
+								 *
+								 * 0x00 Key Pressed 1 time
+								 * 0x01 Key Released
+								 * 0x02 Key Held Down
+								 * 0x03 Key Pressed 2 times
+								 * 0x04 Key Pressed 3 times
+								 * 0x05 Key Pressed 4 times
+								 * 0x06 Key Pressed 5 times
+								 */
+								kaCnt = kA > 0x02? kA - 0x01 : 0x01,
+								cL = cS.toString() + kA.toString(),
+								dS = !_.isEmpty(defaults.metrics.discreteStates) && defaults.metrics.discreteStates[cL]? defaults.metrics.discreteStates[cL] : undefined,
+								st = '',
+								cnt = dS && dS['cnt']? dS['cnt'] : kaCnt,
+								type = dS && dS['type']? dS['type'] : 'B',
+								setAction = function () {
+									switch (kA) {
+										case 0x01:
+											st = dS && dS['action']? dS['action'] : 'release';
+											break;
+										case 0x02:
+											st = dS && dS['action']? dS['action'] : 'hold';
+											break;
+										default:
+											st = dS && dS['action']? dS['action'] : 'press';
+											break;
+									}
+								};
+
+
+							setAction();
+
+							vDev.set("metrics:state", st);
+							vDev.set("metrics:currentScene", cS);
+							vDev.set("metrics:keyAttribute", kA);
+							vDev.set("metrics:maxScenes", mC);
+							vDev.set("metrics:level", cL);
+							vDev.set("metrics:cnt", cnt);
+							vDev.set("metrics:type", type);
+
+						} catch (e) {
+						}
+					}
+				}, "value");
 			}
 		} else if (this.CC["DeviceResetLocally"] === commandClassId) {
 			self.dataBind(self.gateDataBinding, self.zway, nodeId, instanceId, commandClassId, "reset", function(type) {